Основной сервис 1.0 (тесты постман пройдены 1.0)

+++ ewm-service/src/main/java/ru/practicum/model_package/event/dto/EventFullDto.java
@@ -46,7 +46,7 @@ public class EventFullDto {
     //Enum:[ PENDING, PUBLISHED, CANCELED ]
 
 
-    private Integer views; //Количество просмотрев события
+    private Long views; //Количество просмотрев события
 
     private String title; //Заголовок события
 }
